calibration

+++ spund-system/spund_system/lib/ADS_Sensor/ADS_Sensor.cpp
@@ -8,7 +8,7 @@ ADS_Sensor::ADS_Sensor() {}
 
 ADS_Sensor::~ADS_Sensor() {}
 
-void ADS_Sensor::begin(uint8_t address, adsGain_t gain, uint8_t sda, uint8_t scl, uint8_t ads_chan, double offset_vs, uint16_t max_unit)
+void ADS_Sensor::begin(uint8_t address, adsGain_t gain, uint8_t sda, uint8_t scl, uint8_t ads_chan, double offset_vs, double max_unit)
 {
     _ads = new Adafruit_ADS1115();
     Wire.end();
@@ -38,7 +38,7 @@ double ADS_Sensor::readVolts()
 ADS_MA_Meter::ADS_MA_Meter() {}
 ADS_MA_Meter::~ADS_MA_Meter() {}
 
-void ADS_MA_Meter::begin(uint8_t ads_address, adsGain_t ads_gain, uint8_t sda, uint8_t scl, uint8_t ads_chan, double offset_vs, uint8_t max_unit)
+void ADS_MA_Meter::begin(uint8_t ads_address, adsGain_t ads_gain, uint8_t sda, uint8_t scl, uint8_t ads_chan, double offset_vs, double max_unit)
 {
     ADS_Sensor::begin(ads_address, ads_gain, sda, scl, ads_chan, offset_vs, max_unit);
 }
@@ -51,7 +51,7 @@ double ADS_MA_Meter::computeMilliAmps()
 ADS_PH_Meter::ADS_PH_Meter() {}
 ADS_PH_Meter::~ADS_PH_Meter() {}
 
-void ADS_PH_Meter::begin(uint8_t ads_address, adsGain_t ads_gain, uint8_t sda, uint8_t scl, uint8_t ads_chan, double offset_vs, uint8_t max_unit)
+void ADS_PH_Meter::begin(uint8_t ads_address, adsGain_t ads_gain, uint8_t sda, uint8_t scl, uint8_t ads_chan, double offset_vs, double max_unit)
 {
     ADS_Sensor::begin(ads_address, ads_gain, sda, scl, ads_chan, offset_vs, max_unit);
 }
@@ -69,7 +69,7 @@ double ADS_PH_Meter::computeORP()
 ADS_Conductivity_Meter::ADS_Conductivity_Meter() {}
 ADS_Conductivity_Meter::~ADS_Conductivity_Meter() {}
 
-void ADS_Conductivity_Meter::begin(uint8_t ads_address, adsGain_t ads_gain, uint8_t sda, uint8_t scl, uint8_t ads_chan, double offset_vs, uint8_t max_unit)
+void ADS_Conductivity_Meter::begin(uint8_t ads_address, adsGain_t ads_gain, uint8_t sda, uint8_t scl, uint8_t ads_chan, double offset_vs, double max_unit)
 {
     ADS_Sensor::begin(ads_address, ads_gain, sda, scl, ads_chan, offset_vs, max_unit);
 }
@@ -84,7 +84,7 @@ ADS_Pressure_Sensor::ADS_Pressure_Sensor() {}
 
 ADS_Pressure_Sensor::~ADS_Pressure_Sensor() {}
 
-void ADS_Pressure_Sensor::begin(uint8_t ads_address, adsGain_t ads_gain, uint8_t sda, uint8_t scl, uint8_t ads_chan, double offset_vs, uint8_t max_unit)
+void ADS_Pressure_Sensor::begin(uint8_t ads_address, adsGain_t ads_gain, uint8_t sda, uint8_t scl, uint8_t ads_chan, double offset_vs, double max_unit)
 {
     ADS_Sensor::begin(ads_address, ads_gain, sda, scl, ads_chan, offset_vs, max_unit);
 }
@@ -100,7 +100,7 @@ ADS_Level_Sensor::ADS_Level_Sensor() {}
 
 ADS_Level_Sensor::~ADS_Level_Sensor() {}
 
-void ADS_Level_Sensor::begin(uint8_t ads_address, adsGain_t ads_gain, uint8_t sda, uint8_t scl, uint8_t ads_chan, double offset_vs, uint8_t max_unit)
+void ADS_Level_Sensor::begin(uint8_t ads_address, adsGain_t ads_gain, uint8_t sda, uint8_t scl, uint8_t ads_chan, double offset_vs, double max_unit)
 {
     ADS_Sensor::begin(ads_address, ads_gain, sda, scl, ads_chan, offset_vs, max_unit);
 }
